Reorganize repository under RAYEN/MVFA package
