+++ RAYEN/MVFA/CLIP/modified_resnet.py
