+++ RAYEN/MVFA/CLIP/openai.py
