+++ RAYEN/MVFA/train_zero.py
@@ -9,15 +9,15 @@ from torch.nn import functional as F
 from tqdm import tqdm
 from sklearn.metrics import roc_auc_score
 from scipy.ndimage import gaussian_filter
-from dataset.medical_zero import MedTestDataset, MedTrainDataset
-from CLIP.clip import create_model
-from CLIP.tokenizer import tokenize
-from CLIP.adapter import CLIP_Inplanted
+from RAYEN.MVFA.dataset.medical_zero import MedTestDataset, MedTrainDataset
+from RAYEN.MVFA.CLIP.clip import create_model
+from RAYEN.MVFA.CLIP.tokenizer import tokenize
+from RAYEN.MVFA.CLIP.adapter import CLIP_Inplanted
 from PIL import Image
 from sklearn.metrics import precision_recall_curve
-from loss import FocalLoss, BinaryDiceLoss
-from utils import augment, encode_text_with_prompt_ensemble
-from prompt import REAL_NAME
+from RAYEN.MVFA.loss import FocalLoss, BinaryDiceLoss
+from RAYEN.MVFA.utils import augment, encode_text_with_prompt_ensemble
+from RAYEN.MVFA.prompt import REAL_NAME
 
 import warnings
 warnings.filterwarnings("ignore")
